Make improvements based off of feedback in review

+++ server/src/main/java/org/apache/druid/server/coordinator/BalancerStrategy.java
@@ -75,7 +75,7 @@ public interface BalancerStrategy
   BalancerSegmentHolder pickSegmentToMove(
       List<ServerHolder> serverHolders,
       Set<String> broadcastDatasources,
-      int percentOfSegmentsToConsider
+      double percentOfSegmentsToConsider
   );
 
   /**
