+++ server/src/test/java/org/apache/druid/server/coordinator/BalanceSegmentsTest.java
@@ -719,7 +719,7 @@ public class BalanceSegmentsTest
     public BalancerSegmentHolder pickSegmentToMove(
         List<ServerHolder> serverHolders,
         Set<String> broadcastDatasources,
-        int percentOfSegmentsToConsider
+        double percentOfSegmentsToConsider
     )
     {
       return pickOrder.get(pickCounter.getAndIncrement() % pickOrder.size());
